+++ src/runtime/io/Image.hpp
@@ -19,19 +19,19 @@ private:
 
 public:
     using type = Color;
-    using view = unsafe::view2D<type, u16>;
+    using view = unsafe::view2D<type, u32>;
 
 private:
     Owned<type[]> data;
-    u16 width, height;
+    u32 width, height;
 
 public:
     // get data ownership
-    Image(type* data, u16 width, u16 height) :
+    Image(type* data, u32 width, u32 height) :
         data(data), width(width), height(height) {}
 
     [[nodiscard]]
-    auto resize(u16 width, u16 height) -> std::string;
+    auto resize(u32 width, u32 height) -> std::string;
     [[nodiscard]]
     auto write(const std::string_view filename) const -> std::string;
 
@@ -39,7 +39,7 @@ public:
 
     auto swap(Self& other) noexcept -> Self&;
 
-    static auto create(u16 width, u16 height) noexcept -> Self;
+    static auto create(u32 width, u32 height) noexcept -> Self;
 };
 
 NAMESPACE_BEGIN(io)
@@ -49,7 +49,7 @@ NAMESPACE_BEGIN(file)
 auto read_to_image(const std::string_view filename) -> tl::expected<Image, std::string>;
 // hardcode: rgba 8-bit.
 // if error, return error message
-auto write_image(const std::string_view filename, const u8* data, u16 width, u16 height)
+auto write_image(const std::string_view filename, const u8* data, u32 width, u32 height)
     -> std::string;
 
 NAMESPACE_END(file)
